Moving some autoconnector utils into separate package for reuse
