+++ cmd/autoconnector_test.go
