+++ cmd/mixer.go
@@ -14,6 +14,7 @@ import (
 	"time"
 
 	"github.com/jacktrip/jacktrip-agent/pkg/client"
+	"github.com/jacktrip/jacktrip-agent/pkg/common"
 )
 
 // ZitaMode is used to determine the direction the zita service
@@ -359,7 +360,7 @@ func getSampleRateToChannelMap(sentences []string, mode ZitaMode) map[int]int {
 					sampleRates := parseSampleRates(currSentence)
 					// parse the interface's channels
 					r := regexp.MustCompile(`Channels: (\d)`)
-					for ii := j - 1; ii >= client.Max(0, j-5); ii-- {
+					for ii := j - 1; ii >= common.Max(0, j-5); ii-- {
 						currSentence := sentences[ii]
 						subMatch := r.FindStringSubmatch(currSentence)
 						if len(subMatch) > 1 {
